+++ assets/voxygen/shaders/particle-vert.glsl
@@ -75,6 +75,7 @@ const int TORNADO = 33;
 const int DEATH = 34;
 const int ENERGY_BUFFING = 35;
 const int WEB_STRAND = 36;
+const int BLACK_SMOKE = 37;
 
 // meters per second squared (acceleration)
 const float earth_gravity = 9.807;
@@ -191,7 +192,18 @@ void main() {
                     vec3(rand2 * 0.02, rand3 * 0.02, 1.0 + rand4 * 0.1)
                 ),
                 vec3(linear_scale(0.5)),
-                vec4(vec3(0.8, 0.8, 1) * 0.5, start_end(1.0, 0.0)),
+                vec4(vec3(0.8, 0.8, 1) * 0.125 * (3.8 + rand0), start_end(1.0, 0.0)),
+                spin_in_axis(vec3(rand6, rand7, rand8), rand9 * 3 + lifetime * 0.5)
+            );
+            break;
+        case BLACK_SMOKE:
+            attr = Attr(
+                linear_motion(
+                    vec3(0),
+                    vec3(rand2 * 0.02, rand3 * 0.02, 1.0 + rand4 * 0.1)
+                ),
+                vec3(linear_scale(0.5)),
+                vec4(vec3(0.8, 0.8, 1) * 0.125 * (1.8 + rand0), start_end(1.0, 0.0)),
                 spin_in_axis(vec3(rand6, rand7, rand8), rand9 * 3 + lifetime * 0.5)
             );
             break;